--- gobblin-service/src/main/java/org/apache/gobblin/service/modules/orchestration/DagTaskStream.java
@@ -1,88 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.gobblin.service.modules.orchestration;
-
-import java.util.Iterator;
-import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.LinkedBlockingQueue;
-
-import com.google.inject.Singleton;
-
-import lombok.extern.slf4j.Slf4j;
-
-import org.apache.gobblin.annotation.Alpha;
-import org.apache.gobblin.runtime.api.DagActionStore;
-import org.apache.gobblin.runtime.api.MultiActiveLeaseArbiter;
-import org.apache.gobblin.service.modules.orchestration.proc.DagProc;
-import org.apache.gobblin.service.modules.orchestration.task.DagTask;
-import org.apache.gobblin.service.modules.orchestration.task.LaunchDagTask;
-
-
-/**
- * Holds a stream of {@link DagTask}s that {@link DagProcessingEngine} would pull from for processing.
- * Implements {@link Iterator} to provide {@link DagTask}s as soon as it's available to {@link DagProcessingEngine}
- */
-
-@Alpha
-@Slf4j
-@Singleton
-public class DagTaskStream implements Iterator<DagTask>{
-  private final BlockingQueue<DagActionStore.DagAction> dagActionQueue = new LinkedBlockingQueue<>();
-
-  @Override
-  public boolean hasNext() {
-    return !this.dagActionQueue.isEmpty();
-  }
-
-  @Override
-  public DagTask<DagProc> next() {
-    try {
-      DagActionStore.DagAction dagAction = this.dagActionQueue.take();  //`take` blocks till element is not available
-      // todo reconsider the use of MultiActiveLeaseArbiter
-      //MultiActiveLeaseArbiter.LeaseAttemptStatus leaseAttemptStatus = new MultiActiveLeaseArbiter.LeaseObtainedStatus(dagAction);
-      // todo - uncomment after flow trigger handler provides such an api
-      //Properties jobProps = getJobProperties(dagAction);
-      //flowTriggerHandler.getLeaseOnDagAction(jobProps, dagAction, System.currentTimeMillis());
-      //if (leaseAttemptStatus instanceof MultiActiveLeaseArbiter.LeaseObtainedStatus) {
-        // can it return null? is this iterator allowed to return null?
-        return createDagTask(dagAction, new MultiActiveLeaseArbiter.LeaseObtainedStatus(dagAction, System.currentTimeMillis()));
-      //}
-    } catch (Throwable t) {
-      //TODO: need to handle exceptions gracefully
-      log.error("Error getting DagAction from the queue / creating DagTask", t);
-    }
-    return null;
-  }
-
-  private DagTask createDagTask(DagActionStore.DagAction dagAction, MultiActiveLeaseArbiter.LeaseAttemptStatus leaseObtainedStatus) {
-    DagActionStore.FlowActionType flowActionType = dagAction.getFlowActionType();
-
-    switch (flowActionType) {
-      case LAUNCH:
-        return new LaunchDagTask(dagAction, leaseObtainedStatus);
-      default:
-        throw new UnsupportedOperationException("Not yet implemented");
-    }
-  }
-
-  public void addDagAction(DagActionStore.DagAction dagAction) {
-    if (!this.dagActionQueue.offer(dagAction)) {
-      throw new RuntimeException("Could not add dag action " + dagAction + " to the queue");
-    }
-  }
-}
